Reorganized Governance Article
Work towards #339.

Reorganized the article's content and sub-headings for better flow and easier readability.

Merged the section on "Async Governance" into the earlier "Governance Process" section, so all parts related to changing governance can stand alone without any reference to the Governance Meetings parts.

Moved Integrative Election Process out of the section on meetings and into the general section on changing governance (whether in or out of a meeting).

Made sections on electing Circle Rep and Facilitator & Secretary consistent; they now follow a parallel construction and use similar language for similar clauses.

Broadened Process Breakdown language to allow declaring it upon failed governance whether in or out of a meeting (previously it was only allowed for in-meeting failure), but restricted the ability to declare it to Facilitator or Secretary (or Super-Circle Facilitator or Secretary), and excluded Proposer.